% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijdate.R
\name{hijdate}
\alias{hijdate}
\title{Convert Gregorian to Hijri}
\usage{
hijdate(x)
}
\arguments{
\item{x}{a date as a string in dd-mm-yyyy format}
}
\value{
Hijri date as string in dd-mm-yyyy format
}
\description{
Convert Gregorian to Hijri
}
\examples{
x <- "12-12-2021"
hijdate(x)
}
