% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijday.R
\name{hijday}
\alias{hijday}
\title{Get day of the week in Arabic}
\usage{
hijday(x)
}
\arguments{
\item{x}{a date as a string in dd-mm-yyyy format}
}
\value{
Day of the week in Arabic
}
\description{
Get day of the week in Arabic
}
\examples{
x <- "12-12-2021"
hijday(x)
}
