% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijmonth.R
\name{hijmonth}
\alias{hijmonth}
\title{Get name of Islamic month}
\usage{
hijmonth(x)
}
\arguments{
\item{x}{a date as a string in dd-mm-yyyy format}
}
\value{
Arabic name of the month in the Hijri calendar
}
\description{
Get name of Islamic month
}
\examples{
x <- "12-12-2021"
hijmonth(x)
}
